registrarse completo

+++ sitio/src/controllers/usersController.js
@@ -3,7 +3,18 @@ const bcryptjs = require('bcryptjs');//para encriptar el password- que sea brcyp
 const fs = require('fs');
 const {validationResult} = require('express-validator');
 
-//productos
+let usuarios = path.join(__dirname, "../data/users.json");
+let usuariosDB = JSON.parse(fs.readFileSync(usuarios, "utf-8"));
+
+let guardar = (users) => {
+  fs.writeFileSync(
+    path.join(__dirname, "../data/users.json"),
+    JSON.stringify(users, null, " "),
+    "utf-8"
+  );
+};
+
+
 
 const User = require('../models/Users'); 
 
@@ -29,36 +40,33 @@ module.exports = {
       });
     }
     //SI NO ESTA REGISTRADA SIGO CON EL PROCESO, GENERO LA INFORMACION DEL USUARIO
-     let { usuario, nombre, email} = req.body;
-     
+   
      
     if (resultadoValidacion.isEmpty()) {
       let userToCreate = {
-        /* usuario,
+        id: User.generarId(),
+        ...req.body,
         password: bcryptjs.hashSync(req.body.password, 10), 
-        nombre,
-        email,
-        imagenUsuario: req.file.filename, //uso el filename- nombre del archivo de imagen subido
-        rol: "usuario" */
-        ...req.body
+        imagenUsuario:req.file? req.file.filename: 'default-avatar.jpg', //uso el filename- nombre del archivo de imagen subido
+        rol: "usuario"  
       };
-       console.log(userToCreate);
+      
      
+      usuariosDB.push(userToCreate)
+      guardar(usuariosDB) 
 
-     User.crearUsuario(userToCreate); //Creo el usuario con el metodo del modelo
-    console.log(User.crearUsuario(userToCreate));
-
-      return res.redirect('/users/login'); //lo redirige al formulario del login /users/login
+    
+      return res.redirect('/users/login'); 
     } else {
-      //si vienen resultadoValidacion
-      /* if (req.file) {
-        //Borro la imagen subida con multer
+     
+       if (req.file) {
+        
         let imgABorrar = path.join(
           __dirname,
           "../../public/images/users/" + req.file.filename
         );
         fs.unlinkSync(imgABorrar);
-      } */
+      } 
       return res.render("registrarse", {
         errors: resultadoValidacion.mapped(),
         old: req.body,
